backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
websockets==13.1

pdfplumber==0.11.4
pdfminer.six==20231228

pdf2image==1.17.0
pillow==11.0.0
pytesseract==0.3.13

anthropic==0.40.0

openpyxl==3.1.5

httpx==0.28.0
requests==2.32.3

pandas==2.2.3
python-dateutil==2.9.0

python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

aiofiles==24.1.0
